requirements.txt
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
openai>=1.0.0
python-dotenv>=1.0.0
scikit-learn==1.5.2
